use stat from dir info on propfind
When we do a walk on a propfind, we first read the directory entries for
the currently directory prior to recursing into them. If, on recurse, we
failed to read the properties for a subdirectory (for example, because
of permissions), we would just skip that directory entirely.  This lead
to unexpected behaviour, particularly since most file systems show
things in a given directory, regardless of whether they themselves may
be read.

Fix that problem by using the direcetory info we received from the
parent directory when doing the recursion as a fallback in the error
case.